requirements/base.txt
# Latest Django
django==1.10.5

# Configuration
django-environ==0.4.0

# Time zones support
pytz==2015.7

# Task Automation
fabric==1.10.2

# PostgreSQL
psycopg2==2.6.2

# REST API
djangorestframework==3.5.3

# REST Authentication using JSON Web Tokens
djangorestframework-jwt==1.9.0

# Heroku
dj-database-url==0.4.2
